Progress

+++ src/domain/team/repository/TeamRepository.go
@@ -3,5 +3,5 @@ package repository
 import "domain/team/entity"
 
 type TeamRepository interface {
-	FindByIds(ids []uint) []entity.Team
+	FindByIds(ids []uint8) []entity.Team
 }
